backend/requirements.txt
fastapi
uvicorn
transformers
torch
torchvision
pillow
requests
einops
python-multipart